style(lint): Add pre-commit setup and yamllint
In order to support early failings on linting errors during local
development a pre-commit ability was introduced. When making use of it
it will enforce linting rules before every commit so that in the future
less builds will fail because of linting errors. Additionally yamllint
was introduced for local and remote setup. To get a more harmonized code
base in the future. Yamllint configuration was set to defaults that
respect the actual unharmonized code base but will throw warnings to
remind us of the isseu ;) Rules can be made more restrictive in the
future e.g. if no other PRs are open at the same time.