test: add distro as dimension to GH actions matrix (#35)
Until the solution of https://github.com/ansible-community/molecule/issues/3819 it was not possible to consolidate all the molecule platforms. So in the past for each platform there needed to be a seperate entry. Now there is only a default one in the molecule.yml files and the GH Actions expand the distro variable at test time to as many distros as there are configured.